clean up code
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@11419 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genometryImpl/src/com/affymetrix/genometryImpl/util/CachingInputStream.java
@@ -47,7 +47,7 @@ public class CachingInputStream extends FilterInputStream {
 	public int read() throws IOException {
 		byte[] b = new byte[1];
 		int ret = this.read(b, 0, b.length);
-		return (ret == -1 ? ret : ((int) b[0] & 0x00FF));
+		return (ret == -1 ? ret : (b[0] & 0x00FF));
 	}
 
 	@Override
